pyproject.toml
[project]
name = "agentforge-iac"
version = "0.1.0"
description = "Infrastructure as Code for the AgentForge project."
authors = [{ name = "AgentForge Team" }]
requires-python = ">=3.9"
dependencies = [
    "aws-cdk-lib",
    "constructs",
    "pydantic",
]

[project.optional-dependencies]
dev = [
    "ruff",
    "mypy",
    "pre-commit",
    "boto3",
    "boto3-stubs[dynamodb]",
]

[tool.ruff]
line-length = 88
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "C4"]
ignore = ["E501"]

[tool.mypy]
strict = true
ignore_missing_imports = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
exclude = ["build/lib/"]
